backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.0.3

# Database dependencies
sqlalchemy[asyncio]==2.0.23
aiomysql==0.2.0

# AI and computer vision dependencies
opencv-python==4.8.1.78
deepface==0.0.79
tensorflow==2.15.0
keras==2.15.0
scikit-learn==1.3.2
numpy==1.24.4
Pillow==10.1.0

# Geometry processing
shapely==2.0.2

# HTTP client for API calls
aiohttp==3.9.1

# File I/O
aiofiles==23.2.0

# System monitoring
psutil==5.9.6

# Development and utilities
python-dotenv==1.0.0

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 